Refactor shared code
